# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MediaHub API",
    "version": "0.4.0"
  },
  "paths": {
...
    "/api/v1/seedbox/wizard/configure": {
      "post": {
        "summary": "Configure",
        "operationId": "configure_api_v1_seedbox_wizard_configure_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WizardConfiguration"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/seedbox/wizard/advance": {
      "post": {
        "summary": "Advance",
        "operationId": "advance_api_v1_seedbox_wizard_advance_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WizardAdvance"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/seedbox/wizard/vpn": {
      "post": {
        "summary": "Vpn",
        "operationId": "vpn_api_v1_seedbox_wizard_vpn_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VPNImport"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/seedbox/wizard/client": {
      "post": {
        "summary": "Client",
        "operationId": "client_api_v1_seedbox_wizard_client_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientImport"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/seedbox/wizard/preflight": {
      "post": {
        "summary": "Preflight",
        "operationId": "preflight_api_v1_seedbox_wizard_preflight_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WizardExecute"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/seedbox/wizard/install": {
      "post": {
        "summary": "Install",
        "operationId": "install_api_v1_seedbox_wizard_install_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WizardExecute"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/seedbox/wizard/rotate/client": {
      "post": {
        "summary": "Rotate Client",
        "operationId": "rotate_client_api_v1_seedbox_wizard_rotate_client_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientImport"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/seedbox/wizard/rotate/vpn": {
      "post": {
        "summary": "Rotate Vpn",
        "operationId": "rotate_vpn_api_v1_seedbox_wizard_rotate_vpn_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VPNImport"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/seedbox/torrents/add": {
      "post": {
        "summary": "Add",
        "operationId": "add_api_v1_seedbox_torrents_add_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddTorrent"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/seedbox/torrents/action": {
      "post": {
        "summary": "Action",
        "operationId": "action_api_v1_seedbox_torrents_action_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TorrentAction"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/seedbox/locations": {
      "get": {
        "summary": "Locations",
        "operationId": "locations_api_v1_seedbox_locations_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "summary": "Change Location",
        "operationId": "change_location_api_v1_seedbox_locations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VPNLocation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/plex/install-options": {
      "get": {
        "summary": "Options",
        "operationId": "options_api_v1_plex_install_options_get",
        "parameters": [
          {
            "name": "host",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "local",
              "title": "Host"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/plex/install-plan": {
      "post": {
        "summary": "Plan",
        "operationId": "plan_api_v1_plex_install_plan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlexInstallation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/plex/install": {
      "post": {
        "summary": "Install",
        "operationId": "install_api_v1_plex_install_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlexInstallRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "VPNImport": {
        "properties": {
          "revision": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Revision"
          },
          "vpnConfig": {
            "type": "string",
            "maxLength": 65536,
            "minLength": 1,
            "format": "password",
            "title": "Vpnconfig",
            "writeOnly": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "revision",
          "vpnConfig"
        ],
        "title": "VPNImport"
      },
...
      "WizardAdvance": {
        "properties": {
          "revision": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Revision"
          },
          "direction": {
            "type": "string",
            "pattern": "^(next|back)$",
            "title": "Direction"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "revision",
          "direction"
        ],
        "title": "WizardAdvance"
      },
      "WizardConfiguration": {
        "properties": {
          "revision": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Revision"
          },
          "installation": {
            "$ref": "#/components/schemas/SeedboxInstallation"
          },
          "qBittorrent": {
            "$ref": "#/components/schemas/QBitInitialSettings"
          },
          "portForwardingAcknowledged": {
            "type": "boolean",
            "title": "Portforwardingacknowledged",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "revision",
          "installation"
        ],
        "title": "WizardConfiguration"
      },
...
    }
  }
}